Task-3 - Rename of index.js reducer to listApp.js
